// src/features/plants/types/plant.types.ts
export type Plant = {
  id: string;
  nomePopular: string;
  nomeCientifico: string;
  familia: string;
  categoria: string;
  bioma: string;
  origem: string;
  descricao: string;
  caracteristicas: string[];
  usos: string[];
  epocaFloracao: string;
  epocaFrutificacao: string;
  cuidados: string;
  curiosidades: string;
  statusConservacao: string;
  imagem: string;
};